move PhotoDetails to own folder
